assistants/requirements.txt
openai==1.3.7
aws_requests_auth==0.4.3
requests==2.31.0
boto3==1.33.7
pytest-mock==3.12.0
pyyaml==5.4.1
